+++ rust/operator/src/lib.rs
@@ -1,5 +1,7 @@
 mod config;
+mod container;
 mod discovery;
+mod event;
 mod hdfs_controller;
 mod pod_svc_controller;
 mod product_logging;
